Conflicts:
	.gitignore
	modules/core/src/main/java/org/openlmis/core/domain/Right.java
	modules/openlmis-web/build.gradle
	modules/openlmis-web/src/main/webapp/public/js/shared/app.js
	modules/openlmis-web/src/main/webapp/public/pages/navigation.html
	settings.gradle

+++ modules/shipment/src/test/java/org/openlmis/order/repository/mapper/OrderMapperIT.java
@@ -163,7 +163,7 @@ public class OrderMapperIT {
     assertThat(savedOrder.getRnr().getId(), is(expectedOrder.getRnr().getId()));
   }
 
-  private int updateOrderCreatedTime(Order order, Date date) throws SQLException {
+  private long updateOrderCreatedTime(Order order, Date date) throws SQLException {
     List paramList = new ArrayList();
     paramList.add(new java.sql.Date(date.getTime()));
     paramList.add(order.getId());
